requirements.txt
ffmpeg_python==0.2.0
librosa==0.9.2
matplotlib==3.7.1
music21==9.1.0
numpy==1.24.4
piano_transcription_inference==0.0.5
pretty_midi==0.2.10
pytorch_lightning==2.0.4
scikit_learn==1.2.2
torch==2.0.1
tqdm==4.65.0
utils==1.0.1
